Fix random moves in the slide puzzle on stm32
Unfortunately the RNG on the stm32 may fail when called repeatedly in quick successsion - maybe it's running out of entropy.
That would result in the random moves being not random at all because we map the error to 0.

Instead, use rand's pseudo rng in the slide puzzle, seed it once (when it works) and
then re-use the rng throughout.